+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to **DecaLog** are documented in this *changelog*.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and **DecaLog** adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+[Unreleased - Will be 1.12.4]
+### Changed
+- Improved way to handle fatal errors in PHP listener.
+
 ## [1.12.3] - 2020-04-07
 ### Changed
 - Forces mu-plugin dir creation if it doesn't exist.