+++ CHANGELOG.md
@@ -3,10 +3,13 @@ All notable changes to **DecaLog** are documented in this *changelog*.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and **DecaLog** adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-[Unreleased - Will be 1.12.4]
+##[1.12.4] - 2020-04-10
 ### Changed
 - Improved way to handle fatal errors in PHP listener.
-- "Updated user" messages have now "INFO" type (was "notice" previously).
+- "Updated user" messages have now "INFO" type (was "NOTICE" previously).
+- "Logged-in user" messages have now "NOTICE" type (was "INFO" previously).
+### Fixed
+- Some main settings may be not saved.
 
 ## [1.12.3] - 2020-04-07
 ### Changed